Feature/pr template

+++ database/migrations/2026_01_06_011923_create_proveedores_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('proveedores')) {
+            return;
+        }
+
         Schema::create('proveedores', function (Blueprint $table) {
             $table->id();
             $table->foreignId('persona_id')->constrained('personas')->onDelete('cascade');
